requirements.txt
torch
--extra-index-url https://download.pytorch.org/whl/cu118
torch>=2.2.0
matplotlib
numpy
pandas
scipy
seaborn
h5py
biopython
tqdm
fair-esm
gpytorch
hydra-core>=1.3.2